+++ src/components/SidebarBox.js
@@ -1,14 +1,13 @@
 import React from "react";
 import Sidebar from "../containers/Sidebar";
-import Aside from "../components/Aside";
+import Aside from "../style/Aside";
 
 const SidebarBox = () => {
-
-    return(
-        <Aside>
-            <Sidebar />
-        </Aside>
-    )
-}
+  return (
+    <Aside>
+      <Sidebar />
+    </Aside>
+  );
+};
 
 export default SidebarBox;
