+++ src/containers/CallIgdb.js
@@ -1,18 +1,13 @@
 import React, { useState } from "react";
 import axios from "axios";
-import GameCard from "./GameCard";
-import Title from "./Title";
-import Loading from "./Loading";
-import LoadingImg from "./LoadingImg";
-import GameListLayout from "./GameListLayout";
+import Loading from "../style/Loading";
+import Title from "../style/Title";
+import LoadingImg from "../style/LoadingImg";
 
 const API_KEY = process.env.REACT_APP_API_KEY;
 
-const GameListBox = () => {
-  const [gameList, setGameList] = useState([]);
-  const [loading, setLoading] = useState(true);
-
-  if (loading && !gameList.length) {
+const CallIgdb = (props) => {
+  if (props.loading && !props.gameList.length) {
     axios({
       url: "https://cors-anywhere.herokuapp.com/https://api-v3.igdb.com/games",
       method: "POST",
@@ -20,7 +15,7 @@ const GameListBox = () => {
         Accept: "application/json",
         "user-key": API_KEY,
       },
-      data: "fields name, cover.url; limit 20; where total_rating_count>=80;",
+      data: "fields name, cover.url; limit 2; where total_rating_count>=80;",
     })
       .then((response) => response.data)
       .then((data) => {
@@ -32,7 +27,7 @@ const GameListBox = () => {
       });
   }
 
-  if (loading) {
+  if (props.loading) {
     return (
       <Loading>
         <Title>
@@ -45,14 +40,6 @@ const GameListBox = () => {
       </Loading>
     );
   }
-
-  return (
-    <GameListLayout>
-      {gameList.map((item) => (
-        <GameCard {...item} key={item.id} />
-      ))}
-    </GameListLayout>
-  );
 };
 
-export default GameListBox;
+export default CallIgdb;
