+++ src/containers/ProfilPage.js
@@ -3,66 +3,67 @@ import { UserBase } from "../UserBase";
 import axios from "axios";
 import GameCard from "../components/GameCard";
 import Profil from "../components/Profil";
-import ProfilPageLayout from "../components/ProfilPageLayout";
-import ProfilGameLayout from "../components/ProfilGameLayout";
-import ProfilAsideLayout from "../components/ProfilAsideLayout";
-import MyGamovoreLayout from "../components/MyGamovoreLayout";
-import MyGamovoreProfilLayout from "../components/MyGamovoreProfilLayout";
-import SecondaryTitle from "../components/SecondaryTitle";
-import StyleForPseudo from "../components/Pseudo";
-import StyleForAvatar from "../components/Avatar";
-import Loading from "../components/Loading";
-import Title from "../components/Title";
-import LoadingImg from "../components/LoadingImg";
+import ProfilPageLayout from "../style/ProfilPageLayout";
+import ProfilGameLayout from "../style/ProfilGameLayout";
+import ProfilAsideLayout from "../style/ProfilAsideLayout";
+import MyGamovoreLayout from "../style/MyGamovoreLayout";
+import MyGamovoreProfilLayout from "../style/MyGamovoreProfilLayout";
+import SecondaryTitle from "../style/SecondaryTitle";
+import StyleForPseudo from "../style/Pseudo";
+import StyleForAvatar from "../style/Avatar";
+import Loading from "../style/Loading";
+import Title from "../style/Title";
+import LoadingImg from "../style/LoadingImg";
+import CallIgdb from "./CallIgdb";
 
-const API_KEY = process.env.REACT_APP_API_KEY;
+// const API_KEY = process.env.REACT_APP_API_KEY;
 
 const ProfilPage = () => {
-  const [profilGameList, setProfilGameList] = useState([]);
-  const [loading, setLoading] = useState(true);
+     const [gameList, setGameList] = useState([]);
+     const [loading, setLoading] = useState(true);
 
-  if (loading && !profilGameList.length) {
-    axios({
-      url: "https://cors-anywhere.herokuapp.com/https://api-v3.igdb.com/games",
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        "user-key": API_KEY,
-      },
-      data: "fields name, cover.url; limit 2; where total_rating_count>=80;",
-    })
-      .then((response) => response.data)
-      .then((data) => {
-        setProfilGameList(data);
-        setLoading(false);
-      })
-      .catch((err) => {
-        console.error(err);
-      });
-  }
+  //   if (loading && !gameList.length) {
+  //     axios({
+  //       url: "https://cors-anywhere.herokuapp.com/https://api-v3.igdb.com/games",
+  //       method: "POST",
+  //       headers: {
+  //         Accept: "application/json",
+  //         "user-key": API_KEY,
+  //       },
+  //       data: "fields name, cover.url; limit 2; where total_rating_count>=80;",
+  //     })
+  //       .then((response) => response.data)
+  //       .then((data) => {
+  //         setGameList(data);
+  //         setLoading(false);
+  //       })
+  //       .catch((err) => {
+  //         console.error(err);
+  //       });
+  //   }
 
-  if (loading) {
-    return (
-      <Loading>
-        <Title>
-          Be patient young Gamovore, the duck is fishing a games for you ...
-        </Title>
-        <LoadingImg
-          src="https://cdn.dribbble.com/users/591610/screenshots/3861704/pato.gif"
-          alt="loading"
-        />
-      </Loading>
-    );
-  }
-
+  //   if (loading) {
+  //     return (
+  //       <Loading>
+  //         <Title>
+  //           Be patient young Gamovore, the duck is fishing a games for you ...
+  //         </Title>
+  //         <LoadingImg
+  //           src="https://cdn.dribbble.com/users/591610/screenshots/3861704/pato.gif"
+  //           alt="loading"
+  //         />
+  //       </Loading>
+  //     );
+  //   }
+  
   return (
     <ProfilPageLayout>
       <Profil />
       <ProfilAsideLayout>
         <section>
           <SecondaryTitle>My Games</SecondaryTitle>
-          <ProfilGameLayout>
-            {profilGameList.map((item) => (
+          <ProfilGameLayout />
+            {gameList.map((item) => (
               <GameCard little {...item} key={item.id} />
             ))}
           </ProfilGameLayout>
