+++ src/style/Avatar.js
