+++ src/style/Footer.js
