+++ src/style/FooterSpan.js
