+++ src/style/GameCardJacquette.js
