+++ src/style/GameCardName.js
