+++ src/style/GameCardStyle.js
