+++ src/style/GameListLayout.js
