+++ src/style/Linked.js
