+++ src/style/MyGamovoreLayout.js
