+++ src/style/MyGamovoreProfilLayout.js
