+++ src/style/ProfilGameLayout.js
