+++ src/style/Pseudo.js
